refactor(recipe/list): move RecipeListScreen to impl, delegate via BlocScreen
RecipeListScreen and RecipeListPreviews move to
recipe/list/impl/ui. RecipeListItem, RecipeListTestTags,
RecipeFilterOption, and RecipeSortOption stay in public —
they're data/test-tag helpers used by both robots and screens.

RecipeListBloc extends BlocScreen; RecipeListBlocImpl overrides
Content(). BottomNavBloc.Child.RecipeList delegates.

RecipeCategoryFilterScreenshotTest is updated alongside
RecipeListScreenshotTest: imports retarget to impl/ui (including
the SortFilterSheetContent helper used by category-filter tests).

recipe/list/public sets publicResClass = true; recipe/list/impl
gains Compose plugin + ui/text deps.
