+++ client/recipe/list/impl/build.gradle.kts
@@ -1,15 +1,22 @@
-plugins { alias(libs.plugins.kmpLibrary) }
+plugins {
+    alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.compose)
+}
 
 kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(projects.client.recipe.list.public)
+            implementation(projects.client.text.public)
+            implementation(projects.client.ui.public)
             implementation(libs.arkivanov.decompose.core)
+            implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(projects.client.shared)
             implementation(projects.client.recipe.data.public)
             implementation(projects.client.cook.public)
             implementation(projects.client.util.public)
             implementation(libs.multiplatform.settings)
+            implementation(compose.components.resources)
         }
         commonTest.dependencies { implementation(projects.client.recipe.data.testing) }
     }
