+++ client/ui/screenshot-test/build.gradle.kts
@@ -44,6 +44,7 @@ dependencies {
     implementation(project(":client:grocery:core:impl"))
     implementation(project(":client:recipe:core:public"))
     implementation(project(":client:recipe:list:public"))
+    implementation(project(":client:recipe:list:impl"))
     implementation(project(":client:recipe:data:public"))
     implementation(project(":client:settings:public"))
     implementation(project(":client:settings:impl"))
